requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
urllib3==1.26.18